requirements.txt
telethon
jsonlines
python-dotenv